requirements.txt
pyyaml~=6.0
yt_dlp~=2023.7.6
pyarr~=5.2.0
requests~=2.28.1
jellyfin-apiclient-python~=1.10.0
colorlog~=6.8.2
python-dotenv~=1.0.1
themoviedb~=1.0.2